ResourceManager分離
