+++ samples/DXLibrary/basic/SSPlayer/SS5Player.cpp
@@ -2,6 +2,7 @@
 //  SS5Player.cpp
 //
 #include "SS5Player.h"
+#include "SS5ResourceManager.h"
 #include "common/SS5PlayerLibs/SS5PlayerData.h"
 #include "common/SS5PlayerLibs/SS5PlayerTypes.h"
 #include "common/SS5PlayerLibs/SS5PlayerToPointer.h"
@@ -10,71 +11,13 @@
 #include "common/SS5PlayerLibs/SS5AnimCache.h"
 #include "common/SS5PlayerLibs/SS5EffectCache.h"
 #include "common/SS5PlayerLibs/SS5ResourceSet.h"
+#include "common/Helper/Util.h"
 #include "common/Animator/ssplayer_matrix.h"
 #include "common/Animator/ssplayer_macro.h"
 
 namespace ss
 {
 
-/**
- * definition
- */
-
-static const ss_u32 DATA_ID = 0x42505353;
-static const ss_u32 DATA_VERSION = 4;
-
-
-/**
- * utilites
- */
-static void splitPath(std::string& directoty, std::string& filename, const std::string& path)
-{
-    std::string f = path;
-    std::string d = "";
-
-    size_t pos = path.find_last_of("/");
-	if (pos == std::string::npos) pos = path.find_last_of("\\");	// for win
-
-    if (pos != std::string::npos)
-    {
-        d = path.substr(0, pos+1);
-        f = path.substr(pos+1);
-    }
-
-	directoty = d;
-	filename = f;
-}
-
-// printf 形式のフォーマット
-#ifndef va_copy
-#    define va_copy(dest, src) ((dest) = (src))
-#endif
-static std::string Format(const char* format, ...){
-
-	static std::vector<char> tmp(1000);
-
-	va_list args, source;
-	va_start(args, format);
-	va_copy( source , args );
-
-	while (1)
-	{
-		va_copy( args , source );
-		//Windows
-		if (_vsnprintf(&tmp[0], tmp.size(), format, args) == -1)
-		{
-			tmp.resize(tmp.size() * 2);
-		}
-		else
-		{
-			break;
-		}
-	}
-	tmp.push_back('\0');
-	std::string ret = &(tmp[0]);
-	va_end(args);
-	return ret;
-}
 
 //座標回転処理
 //指定した座標を中心に回転後した座標を取得します
@@ -105,226 +48,6 @@ unsigned int getRandomSeed()
 }
 
 
-
-/**
- * ResourceManager
- */
-
-static ResourceManager* defaultInstance = NULL;
-const std::string ResourceManager::s_null;
-
-ResourceManager* ResourceManager::getInstance()
-{
-	if (!defaultInstance)
-	{
-		defaultInstance = ResourceManager::create();
-	}
-	return defaultInstance;
-}
-
-ResourceManager::ResourceManager(void)
-{
-}
-
-ResourceManager::~ResourceManager()
-{
-	removeAllData();
-}
-
-ResourceManager* ResourceManager::create()
-{
-	ResourceManager* obj = new ResourceManager();
-	return obj;
-}
-
-ResourceSet* ResourceManager::getData(const std::string& dataKey)
-{
-	ResourceSet* rs = _dataDic.at(dataKey);
-	return rs;
-}
-
-std::string ResourceManager::addData(const std::string& dataKey, const ProjectData* data, const std::string& imageBaseDir)
-{
-	SS_ASSERT2(data != NULL, "Invalid data");
-	SS_ASSERT2(data->dataId == DATA_ID, "Not data id matched");
-	SS_ASSERT2(data->version == DATA_VERSION, "Version number of data does not match");
-	
-	// imageBaseDirの指定がないときコンバート時に指定されたパスを使用する
-	std::string baseDir = imageBaseDir;
-	if (imageBaseDir == s_null && data->imageBaseDir)
-	{
-		ToPointer ptr(data);
-		const char* dir = ptr.toString(data->imageBaseDir);
-		baseDir = dir;
-	}
-
-	//アニメはエフェクトを参照し、エフェクトはセルを参照するのでこの順番で生成する必要がある
-	CellCache* cellCache = CellCache::create(data, baseDir);
-
-	EffectCache* effectCache = EffectCache::create(data, baseDir, cellCache);	//
-
-	AnimeCache* animeCache = AnimeCache::create(data);
-
-	ResourceSet* rs = new ResourceSet();
-	rs->data = data;
-	rs->isDataAutoRelease = false;
-	rs->cellCache = cellCache;
-	rs->animeCache = animeCache;
-	rs->effectCache = effectCache;
-	_dataDic.insert(std::map<std::string, ResourceSet*>::value_type(dataKey, rs));
-
-	return dataKey;
-}
-
-std::string ResourceManager::addDataWithKey(const std::string& dataKey, const std::string& ssbpFilepath, const std::string& imageBaseDir)
-{
-
-	std::string fullpath = ssbpFilepath;
-
-	unsigned long nSize = 0;
-	void* loadData = SSFileOpen(fullpath.c_str(), "rb", &nSize);
-	if (loadData == NULL)
-	{
-		std::string msg = "Can't load project data > " + fullpath;
-		SS_ASSERT2(loadData != NULL, msg.c_str());
-	}
-	
-	const ProjectData* data = static_cast<const ProjectData*>(loadData);
-	SS_ASSERT2(data->dataId == DATA_ID, "Not data id matched");
-	SS_ASSERT2(data->version == DATA_VERSION, "Version number of data does not match");
-	
-	std::string baseDir = imageBaseDir;
-	if (imageBaseDir == s_null)
-	{
-		// imageBaseDirの指定がないとき
-		if (data->imageBaseDir)
-		{
-			// コンバート時に指定されたパスを使用する
-			ToPointer ptr(data);
-			const char* dir = ptr.toString(data->imageBaseDir);
-			baseDir = dir;
-		}
-		else
-		{
-			// プロジェクトファイルと同じディレクトリを指定する
-			std::string directory;
-			std::string filename;
-			splitPath(directory, filename, ssbpFilepath);
-			baseDir = directory;
-		}
-		//SSLOG("imageBaseDir: %s", baseDir.c_str());
-	}
-
-	addData(dataKey, data, baseDir);
-	
-	// リソースが破棄されるとき一緒にロードしたデータも破棄する
-	ResourceSet* rs = getData(dataKey);
-	SS_ASSERT2(rs != NULL, "");
-	rs->isDataAutoRelease = true;
-	
-	return dataKey;
-}
-
-std::string ResourceManager::addData(const std::string& ssbpFilepath, const std::string& imageBaseDir)
-{
-	// ファイル名を取り出す
-	std::string directory;
-    std::string filename;
-	splitPath(directory, filename, ssbpFilepath);
-	
-	// 拡張子を取る
-	std::string dataKey = filename;
-	size_t pos = filename.find_last_of(".");
-    if (pos != std::string::npos)
-    {
-        dataKey = filename.substr(0, pos);
-    }
-
-	//登録されている名前か判定する
-	std::map<std::string, ResourceSet*>::iterator it = _dataDic.find(dataKey);
-	if (it != _dataDic.end())
-	{
-		//登録されている場合は処理を行わない
-		std::string str = "";
-		return str;
-	}
-
-	return addDataWithKey(dataKey, ssbpFilepath, imageBaseDir);
-}
-
-void ResourceManager::removeData(const std::string& dataKey)
-{
-	ResourceSet* rs = getData(dataKey);
-
-	//テクスチャの解放
-	releseTexture(dataKey);
-	
-	//バイナリデータの削除
-	delete rs;
-	_dataDic.erase(dataKey);
-}
-
-void ResourceManager::removeAllData()
-{
-	//全リソースの解放
-	while (!_dataDic.empty())
-	{
-		std::map<std::string, ResourceSet*>::iterator it = _dataDic.begin();
-		std::string ssbpName = it->first;
-		removeData(ssbpName);
-	}
-	_dataDic.clear();
-}
-
-//データ名、セル名を指定して、セルで使用しているテクスチャを変更する
-bool ResourceManager::changeTexture(char* ssbpName, char* ssceName, long texture)
-{
-	bool rc = false;
-
-	ResourceSet* rs = getData(ssbpName);
-	rc = rs->cellCache->setCellRefTexture(rs->data, ssceName, texture);
-
-	return(rc);
-}
-
-//指定したデータのテクスチャを破棄します
-bool ResourceManager::releseTexture(std::string ssbpName)
-{
-
-	ResourceSet* rs = getData(ssbpName);
-	bool rc = rs->cellCache->releseTexture(rs->data);
-
-	return(rc);
-}
-
-//アニメーションの総フレーム数を取得する
-int ResourceManager::getMaxFrame(std::string ssbpName, std::string animeName)
-{
-	int rc = -1;
-
-	ResourceSet* rs = getData(ssbpName);
-	AnimeRef* animeRef = rs->animeCache->getReference(animeName);
-	if (animeRef == NULL)
-	{
-		std::string msg = Format("Not found animation > anime=%s", animeName.c_str());
-		SS_ASSERT2(animeRef != NULL, msg.c_str());
-	}
-	rc = animeRef->animationData->numFrames;
-
-	return(rc);
-}
-
-//ssbpファイルが登録されているかを調べる
-bool ResourceManager::isDataKeyExists(const std::string& dataKey) {
-	// 登録されている名前か判定する
-	std::map<std::string, ResourceSet*>::iterator it = _dataDic.find(dataKey);
-	if (it != _dataDic.end()) {
-		//登録されている
-		return true;
-	}
-
-	return false;
-}
 
 /**
  * Player
