+++ samples/DXLibrary/basic/SSPlayer/SS5Player.h
@@ -105,36 +105,11 @@ class ResourceSet;
 struct ProjectData;
 class SSSize;
 class Player;
+class ResourceManager;
 
 //関数定義
 extern void get_uv_rotation(float *u, float *v, float cu, float cv, float deg);
 
-/**
-* 定数
-*/
-
-#define __SSPI__	(3.14159265358979323846f)
-#define __SS2PI__	(__SSPI__ * 2)
-
-#define SS_SAFE_DELETE(p)            do { if(p) { delete (p); (p) = 0; } } while(0)
-#define SS_SAFE_DELETE_ARRAY(p)     do { if(p) { delete[] (p); (p) = 0; } } while(0)
-#define SS_SAFE_FREE(p)                do { if(p) { free(p); (p) = 0; } } while(0)
-#define SS_SAFE_RELEASE(p)            do { if(p) { (p)->release(); } } while(0)
-#define SS_SAFE_RELEASE_NULL(p)        do { if(p) { (p)->release(); (p) = 0; } } while(0)
-#define SS_SAFE_RETAIN(p)            do { if(p) { (p)->retain(); } } while(0)
-#define SS_BREAK_IF(cond)            if(cond) break
-
-#ifdef _DEBUG
-	#define SSLOG(...)       do {} while (0)
-	#define SS_ASSERT(cond)    assert(cond)
-	#define SS_ASSERT2(cond, msg) SS_ASSERT(cond)
-	#define SSLOGERROR(format,...)  do {} while (0)
-#else
-	#define SSLOG(...)       do {} while (0)
-	#define SS_ASSERT(cond)
-	#define SS_ASSERT2(cond, msg) ((void)(cond))
-	#define SSLOGERROR(format,...)  do {} while (0)
-#endif
 
 
 /**
@@ -287,117 +262,6 @@ public:
 public:
 };
 
-
-/**
- * ResourceManager
- */
-//class ResourceManager : public Ref
-class ResourceManager
-{
-public:
-	static const std::string s_null;
-
-	/**
-	 * デフォルトインスタンスを取得します.
-	 *
-	 * @return デフォルトのResourceManagerインスタンス
-	 */
-	static ResourceManager* getInstance();
-
-	/**
-	 * ssbpファイルを読み込み管理対象とします.
-	 * dataKeyはssbpのファイル名（拡張子なし）になります.
-	 *
-	 * @param  ssbpFilepath  ssbpファイルのパス
-	 * @param  imageBaseDir  画像ファイルの読み込み元ルートパス. 省略時はssbpのある場所をルートとします.
-	 * @return dataKey
-	 */
-	std::string addData(const std::string& ssbpFilepath, const std::string& imageBaseDir = s_null);
-
-	/**
-	 * ssbpファイルを読み込み管理対象とします.
-	 *
-	 * @param  dataKey       dataKeyの指定
-	 * @param  ssbpFilepath  ssbpファイルのパス
-	 * @param  imageBaseDir  画像ファイルの読み込み元ルートパス. 省略時はssbpのある場所をルートとします.
-	 * @return dataKey
-	 */
-	std::string addDataWithKey(const std::string& dataKey, const std::string& ssbpFilepath, const std::string& imageBaseDir = s_null);
-
-	/**
-	 * 指定されたssbpデータを管理対象とします.
-	 *
-	 * @param  dataKey       dataKeyの指定
-	 * @param  data          ssbpデータ
-	 * @param  imageBaseDir  画像ファイルの読み込み元ルートパス. 省略時はssbpのある場所をルートとします.
-	 * @return dataKey
-	 */
-	std::string addData(const std::string& dataKey, const ProjectData* data, const std::string& imageBaseDir = s_null);
-	
-	/**
-	 * 指定データを解放します.
-	 * パス、拡張子を除いたssbp名を指定してください。
-	 *
-	 * @param  dataKey
-	 */
-	void removeData(const std::string& dataKey);
-
-	/**
-	 * 全てのデータを解放します.
-	 */
-	void removeAllData();
-
-	/**
-	* 名前に対応するデータ取得します.
-	*/
-	ResourceSet* getData(const std::string& dataKey);
-
-	/**
-	* 指定したセルのテクスチャを変更します.
-	* @param  ssbpName       ssbp名（拡張子を除くファイル名）
-	* @param  ssceName       ssce名（拡張子を除くファイル名）
-	* @param  texture        変更後のテクスチャハンドル
-	* @return 変更を行ったか
-	*/
-	bool changeTexture(char* ssbpName, char* ssceName, long texture);
-
-	/**
-	* 指定したデータのテクスチャを破棄します。
-	* @param  dataName       ssbp名（拡張子を除くファイル名）
-	* @return 成功失敗
-	*/
-	bool releseTexture(std::string);
-
-	/**
-	* 読み込んでいるssbpからアニメーションの総フレーム数を取得します。
-	* @param  ssbpName       ssbp名（拡張子を除くファイル名）
-	* @param  animeName      ssae/モーション名
-	* @return アニメーションの総フレーム（存在しない場合はアサート）
-	*/
-	int getMaxFrame(std::string ssbpName, std::string animeName);
-
-	/**
-	* 名前が登録されていればtrueを返します
-	*
-	* @param dataKey
-	* @return
-	*/
-	bool isDataKeyExists(const std::string& dataKey);
-
-	/**
-	 * 新たなResourceManagerインスタンスを構築します.
-	 *
-	 * @return ResourceManagerインスタンス
-	 */
-	static ResourceManager* create();
-
-public:
-	ResourceManager(void);
-	virtual ~ResourceManager();
-
-protected:
-	std::map<std::string, ResourceSet*>	_dataDic;
-};
